refactor: エントリーポイントでヘルパーを切る
