+++ cmd/gateway.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"net/url"
 	"os"
 	"os/signal"
 	"strings"
@@ -63,8 +62,7 @@ func runGateway(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	dsn := fmt.Sprintf(
-		"postgresql://%s:%s@%s:%d/%s",
+	dsn := postgresDSN(
 		cfg.PostgreSQLUser,
 		cfg.PostgreSQLPassword,
 		cfg.PostgreSQLHost,
@@ -176,53 +174,4 @@ func runGateway(cmd *cobra.Command, args []string) error {
 	shutdownCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
 	return srv.Shutdown(shutdownCtx)
-}
-
-type connectRoute struct {
-	path    string
-	handler http.Handler
-}
-
-func route(path string, h http.Handler) connectRoute {
-	return connectRoute{path: path, handler: h}
-}
-
-func mountAll(r chi.Router, routes ...connectRoute) {
-	for _, route := range routes {
-		r.Mount(route.path, route.handler)
-	}
-}
-
-func splitCSV(value string) []string {
-	if value == "" {
-		return nil
-	}
-	parts := strings.Split(value, ",")
-	items := make([]string, 0, len(parts))
-	for _, part := range parts {
-		part = strings.TrimSpace(part)
-		if part == "" {
-			continue
-		}
-		items = append(items, part)
-	}
-	return items
-}
-
-func allowedOrigins(redirects []string) []string {
-	origins := make([]string, 0, len(redirects))
-	seen := map[string]struct{}{}
-	for _, redirect := range redirects {
-		u, err := url.Parse(redirect)
-		if err != nil || u.Scheme == "" || u.Host == "" {
-			continue
-		}
-		origin := u.Scheme + "://" + u.Host
-		if _, ok := seen[origin]; ok {
-			continue
-		}
-		seen[origin] = struct{}{}
-		origins = append(origins, origin)
-	}
-	return origins
 }
